remove eviction module
