+++ src/storage/frame.rs
@@ -2,10 +2,7 @@
 //!
 //! TODO more docs.
 
-use crate::storage::{
-    eviction::EvictionState,
-    frame_group::{FrameGroup, FRAME_GROUP_SIZE},
-};
+use crate::storage::frame_group::{EvictionState, FrameGroup, FRAME_GROUP_SIZE};
 use crate::{
     bpm::BufferPoolManager,
     page::{Page, PAGE_SIZE},
